Categorize functions
filterEmpty()       ->  filter.empty()
is.empty()          ->  is.empty()
is.objectLiteral()  ->  is.objectLiteral()
numberIsNaN()       ->  equals.nan()
shallowEqual()      ->  compare.shallow()
