+++ src/compare/index.ts
@@ -1,0 +1,1 @@
+export { shallow } from "./shallow";
