+++ src/equals/index.ts
@@ -1,0 +1,1 @@
+export { nan } from "./nan";
