+++ src/index.ts
@@ -1,5 +1,11 @@
-export { filterEmpty } from "./filterEmpty";
-export { isEmpty } from "./isEmpty";
-export { isObjectLiteral } from "./isObjectLiteral";
-export { numberIsNaN } from "./numberIsNaN";
-export { shallowEqual } from "./shallowEqual";
+import * as compare from "./compare";
+import * as equals from "./equals";
+import * as filter from "./filter";
+import * as is from "./is";
+
+export {
+    compare,
+    equals,
+    filter,
+    is,
+};
