+++ src/is/index.ts
@@ -1,0 +1,2 @@
+export { empty } from "./empty";
+export { objectLiteral } from "./objectLiteral";
